Improvement: no need, that I know of, to restrict this to FreeStyleProject types.
Bug: Found that this plugin doesn't work properly when two or more
projects depend on the same snapshot dependency project. It was
comparing the snapshot resolution time only if a snapshot was downloaded
which wouldn't happen a second time when using the user's local
repository. I changed that to compare the meta-data's lastUpdated
against the last run date of the polling project.